pyproject.toml
[project]
name = "aftermark"
version = "0.1.1"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pillow", 
]

[project.optional-dependencies]
dev = ["matplotlib", "opencv-python", "scipy"]

[project.scripts] 
nuke = "aftermark.cli.nuke:main"

[build-system]
requires      = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where    = [""]       
include  = ["aftermark*"] 
exclude  = ["demo*", "lsb*", "tests*", "jpg", "artifacts"]
